add better joined query test
